+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sql Server `..` syntax is no longer used (now uses `.dbo.` - or whatever the table schema is).  Since references can be shared by users the default schema notation is not good idea.
 - Cohort Query Bulder will now connect to the database containing the data rather than the users default database when querying data on a single database
 
+### Fixed
+
+- Fixed handling of credentials where password is blank (allowed)
+
 ## [3.2.1] - 2019-10-30
 
 ### Added